requirements.txt
Django==2.2.4
psycopg2==2.8.3
djangorestframework==3.10.2
djangorestframework-jwt==1.10.0
django-cors-headers==3.1.0